make a more generic implementation
