+++ src/blocks/models.rs
@@ -10,9 +10,3 @@ pub enum BlockContent {
     SNARKJobs,
     Analytics,
 }
-
-pub struct BlockMultiSearch {
-    pub block_height: Option<i64>,
-    pub public_key: Option<String>,
-    pub state_hash: Option<String>,
-}
